Retry trashbin in case of timestamp collision (#730)
In OC storage the file name pattern for deleted files include the
timestamp: "deleted.d12345678". In some rare cases like when running
tests, it can happen that the deletion happens so fast that it causes
timestamp collisions for the same file names.

To fix this, if we detect the collision case we retry again with a
higher timestamp value.