Refactor AppProvider workflow and protocol (#1035)
* Refactored the OpenFileInAppProvider workflow

This included changing the gRPC protocol and removing
storageID and UIURL from config.

En passant, this fixes #991 and includes other minor changes
concerning error handling and logging.

* Linting and other fixes + changelog

* Added configuration for ODF and MD files

* Fixed error handling + improved logging of a stack trace

* Further fixes following tests

* Improved URL handling following review

Also cleared some TODOs and reduced HTTP timeout to 5s

* Refactored the /wopi/cbox/endpoints HTTP call

This is now in its own function, as it does not need to be executed
at each incoming request. Added a TODO for implementing a refresh
every day or so.

* Do not use for now a shared map.

TODO for a future PR: refresh the map of apps URLs every day or week,
and cache it at the service level (protected by a multi-reader Lock).