Handle race conditions when listing xattrs better
The read lock didn't prevent running into a rare but conistent error when
listing the extended attributes of a node. A much more robust solution
seems to be to just retry once when listing didn't work initially.
I haven't been able to get the unit tests to flake with this change
anymore.