Add tracing information to events.
This adds a TraceParent field to the event envelope, adds context as
a parameter to the Publish function, adds functionality to extract
the traceparent from the context, and added a method to the envelope
to recover the tracing info and add it to the context.

This also passes context (empty if no context was present in the calling
function) to Publish wherever it is called.