fix(auth/scope): fix nested resource check when creating a new resource
When creating a reource (e.g a document via the app/new endpoint) below a next
folder structure of a public link, we can't stat the resource itself (it
doesn't exit yet) for checking if it is a descendant of the share root.
We now stat the resource's parent instead in that case.

Fixes: https://github.com/owncloud/ocis/issues/8957