fix: quotes in content-disposition header
Replaces the quotes by actually encoding the filename in the dav `Content-Disposition` header. The value of the `filename*` parameter must not be surrounded by any quotes, but rather be encoded in the first place. See RFC-6266 for more details.

The quotes caused an issue where certain browsers would decode the quotes and falsely prepend them to the filename.