Do not send "delete" sses when items are moved
They trigger an item-renamed event instead if the file is moved to a
different location in the same space or a delete if the file is moved
out of the OC storage root.
The only problematic case is when the file is moved between spaces, in
this case there won't be a proper SSE and a page refresh is required to
make it disappear from the file list.