fix(users): consitently fetch the correct LDAP User Attributes
The list of requested Attributes was duplicated and missed the newly
introduced tenantId Attribute in some place. This introduces a helper
to get the correct list of attribute where ever that is required.

Also: Don't request attributes for which now schema mapping is defined,
as that looks ugly on the server logs at might even confuse servers.