fix: revision handling for parallel uploads to the same node
When multiple uploads on the same node are in post-processing at the
same time the upload that enters post-processing 2nd will create a
new file revision where the 1st upload is supposed to endup.
However if the upload finished post-processing before the 1st upload
the blob of the first upload ended up over-writing the original node.
We now make sure that the 1st upload is writing the blob to the revision
node.

Closes: #1867