fix: restore revision after negative post-processing outcome
When the postprocessing for a file results in e.g. the
"PPOutcomeDelete" state it is not enough to just restore the metadata.

On posixfs we also need to restore the file blob itself as another
concurrent upload might have written it's blob already into the revison
node that was created during the postprocessing of the to be reverted
upload.

Closes #1871